+++ libraries/zf_device/zf_device_oled.h
@@ -104,7 +104,7 @@ void    oled_draw_point                 (uint16 x, uint16 y, const uint8 color);
 void    oled_show_string                (uint16 x, uint16 y, const char ch[]);
 void    oled_show_int                   (uint16 x, uint16 y, const int32 dat, uint8 num);
 void    oled_show_uint                  (uint16 x, uint16 y, const uint32 dat, uint8 num);
-void    oled_show_float                 (uint16 x, uint16 y, const float dat, uint8 num, uint8 pointnum);
+void    oled_show_float                 (uint16 x, uint16 y, const double dat, uint8 num, uint8 pointnum);
 
 void    oled_show_binary_image          (uint16 x, uint16 y, const uint8 *image, uint16 width, uint16 height, uint16 dis_width, uint16 dis_height);
 void    oled_show_gray_image            (uint16 x, uint16 y, const uint8 *image, uint16 width, uint16 height, uint16 dis_width, uint16 dis_height, uint8 threshold);
